完善多选功能

+++ app/src/main/java/remix/myplayer/model/Genre.java
@@ -7,7 +7,7 @@ package remix.myplayer.model;
  * @Date 2016/9/7 09:37
  */
 public class Genre {
-    public long GenreID;
+    public int GenreID;
     public String GenreName;
     public Genre(int id,String name){
         GenreID = id;
